+++ test/clj_cbor/core_test.clj
@@ -29,12 +29,16 @@
     (is (= {Long :y} (:write-handlers codec)))))
 
 
+(deftest multi-value-coding
+  (is (= [true 123 :abc] (cbor/decode-seq (cbor/encode-seq [true 123 :abc])))))
+
+
 (deftest eof-handling
   (testing "sequence of values"
     (is (= (list :a 123 true "foo") (decode-hex-all cbor/default-codec "D827623A61187BF563666F6F"))))
   (testing "interrupted data"
     (is (cbor-error? :clj-cbor.codec/end-of-input
-          (decode-hex "D827623A61187BF563666F")))))
+          (decode-hex-all "D827623A61187BF563666F")))))
 
 
 (deftest slobber-utils
@@ -45,4 +49,7 @@
     (is (= 21 (cbor/spit file data-b :append true)))
     (is (= 81 (.length file)))
     (is (= data-a (cbor/slurp file)))
-    (is (= [data-a data-b] (cbor/slurp-all file)))))
+    (is (= [data-a data-b] (cbor/slurp-all file)))
+    (.delete file)
+    (is (= 81 (cbor/spit-all file [data-b data-a])))
+    (is (= [data-b data-a] (cbor/slurp-all file)))))